combine two existing Modal component, and update  all the references
