--- src/components/ModalTest/Modal.module.scss
@@ -1,60 +1,0 @@
-.modal_container{
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  right: auto;
-  bottom: auto;
-  width: auto;
-  max-width: 500px;
-  max-height: 90%;
-  border-radius: 5px;
-  transform: translate(-50%, -50%);
-  background-color: white;
-  box-shadow: rgba(0, 0, 0, 0.5) 0px 0px 10px;
-  
-  .close_button_wrapper{
-    position: absolute;
-    top: 0;
-    right: 0;
-    height: 60px;
-    display: flex;
-    align-items: center;
-    padding-right: 20px;
-    z-index: 1;
-  }
-
-  %center{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .header{
-    @extend %center;
-    flex-grow: 1;
-    min-height: 60px;
-    padding: 0 44px;
-    color: #292b32;
-    font-family: museo-regular, "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
-    font-size: 18px;
-    line-height: 24px;
-    letter-spacing: 0.15px;
-    text-align: center;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    overflow: hidden;
-  }
-
-  .content{
-    flex-grow: 1;
-    max-height: 400px;
-    overflow: hidden auto;
-    padding: 10px 20px;
-  }
-
-  .footer{
-    @extend %center;
-    padding: 10px 20px;
-  }
-}
-
