+++ src/pages/PostTask/AlertModal/AlertModal.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import styles from './AlertModal.module.scss';
 
 import Button from '../../../components/Button';
-import Modal from '../../../components/ModalTest';
+import Modal from '../../../components/Modal';
 
 export default function AlertModal({ onLeftBtnClick, onRightBtnClick, onRequestClose }) {
   const title = 'Sorry to see you go...';
